created AirConditioner entity, moved Optionals from its former and obsolete package to the new one, also moved TractionControl
