--- src/main/java/com/develhope/spring/model/vehicle/optionals/Optionals.java
@@ -1,5 +1,0 @@
-package com.develhope.spring.model.vehicle.optionals;
-
-public interface Optionals {
-
-}
