+++ src/main/java/com/develhope/spring/vehicle/optionals/models/Optionals.java
@@ -1,0 +1,5 @@
+package com.develhope.spring.vehicle.optionals.models;
+
+public interface Optionals {
+
+}
